// src/api/types.ts
export interface ValidationItem {
  id: number
  decision: string
  validateur: string
  dateValidation: string
  motif: string | null
}

export interface DemandeConge {
  id: number
  reference: string
  employeId: number
  employeNom: string
  employeInitiales: string
  departement: string | null
  typeCongeId: number
  typeNom: string
  typeCode: string | null
  couleur: string | null
  dateDebut: string
  dateFin: string
  nombreJours: number
  motif: string | null
  statut: 'EN_ATTENTE' | 'APPROUVEE' | 'REFUSEE' | 'ANNULEE'
  motifRefus: string | null
  justificatifUrl: string | null
  dateDemande: string
  dateValidation: string | null
  validePar: string | null
  historique: ValidationItem[]
}
// ...
export interface AbsenceMensuelle {
  mois: string
  jours: number
}

export interface DeptCount {
  name: string
  value: number
}
// ...
export type StatutFrais = 'EN_ATTENTE' | 'EN_COURS' | 'REMBOURSEE' | 'REFUSEE' | 'ANNULEE'

export interface NoteFrais {
  id: number
  reference: string
  employeId: number
  employeNom: string
  employeInitiales: string
  departement: string | null
  titre: string
  devise: string
  date: string
  priorite: string
  remarque: string | null
  montantTotal: number
  statut: StatutFrais
  motifRefus: string | null
  dateCreation: string
  nbDepenses: number
  depenses: string[]
  justificatifs: string[]
}
// ...
export interface Candidature {
  id: number
  offreId: number
  offreTitre: string
  candidatId: number
  candidatNom: string
  candidatPrenom: string
  nomComplet: string
  initiales: string
  email: string
  telephone: string | null
  linkedin: string | null
  etape: 'BOITE_RECEPTION' | 'BROUILLON' | 'ENTRETIEN_TEL' | 'ENTRETIEN_PHYSIQUE' | 'EMBAUCHE'
  dateEntretien: string | null
  dateCreation: string
  cvDisponible: boolean
  cvNom: string | null
  lettreDisponible: boolean
  historique: string[]
}